requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
uvicorn>=0.34.0,<0.35.0
psycopg2-binary>=2.9.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.7.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0
asyncpg>=0.30.0,<0.32.0
pydantic-settings>=2.0.0,<3.0.0
sqlmodel>=0.0.33
passlib>=1.7.0,<2.0.0
bcrypt>=3.2.0,<4.0.0
pyjwt>=2.0.0,<3.0.0